+++ apps/rsbuild/ssr/css/src/server.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @nx/enforce-module-boundaries
 import { createServer } from '@ng-rsbuild/plugin-angular/ssr';
 import bootstrap from './main.server';