+++ packages/compiler/mocks/fixtures/integration/minimal/rsbuild.mock.config.ts
@@ -1,0 +1,7 @@
+import { createConfig } from '../../../../src';
+
+export default createConfig({
+  browser: './src/main.ts',
+  inlineStylesExtension: 'css',
+  tsconfigPath: './mocks/fixtures/integration/minimal/tsconfig.mock.json',
+});