+++ packages/compiler/src/compilation/index.ts
@@ -1,0 +1,5 @@
+export * from './augments';
+export * from './build-and-analyze';
+export * from './file-emitter';
+export * from './setup-compilation';
+export * from './setup-with-paralell-compilation';