+++ packages/rsbuild-plugin-angular/CHANGELOG.md
@@ -1,3 +1,55 @@
+## 19.0.0-alpha.24 (2025-01-24)
+
+This was a version bump only for rsbuild-plugin-angular to align it with other projects, there were no code changes.
+
+## 19.0.0-alpha.23 (2025-01-24)
+
+### 🩹 Fixes
+
+- **rsbuild-plugin-angular:** close parallel compilation workers ([#76](https://github.com/Coly010/ng-rspack-build/pull/76))
+
+### ❤️ Thank You
+
+- Edouard Bozon
+
+## 19.0.0-alpha.22 (2025-01-23)
+
+This was a version bump only for rsbuild-plugin-angular to align it with other projects, there were no code changes.
+
+## 19.0.0-alpha.21 (2025-01-23)
+
+This was a version bump only for rsbuild-plugin-angular to align it with other projects, there were no code changes.
+
+## 19.0.0-alpha.20 (2025-01-22)
+
+### 🚀 Features
+
+- **compiler:** add compiler package to house compiler logic ([#63](https://github.com/Coly010/ng-rspack-build/pull/63))
+- **rsbuild-plugin-angular:** add withConfigurations and fileReplacement support #43 ([#46](https://github.com/Coly010/ng-rspack-build/pull/46), [#43](https://github.com/Coly010/ng-rspack-build/issues/43))
+- **rsbuild-plugin-angular:** allow overriding paths to static assets ([2c6f318](https://github.com/Coly010/ng-rspack-build/commit/2c6f318))
+- **docs:** add docs site ([#42](https://github.com/Coly010/ng-rspack-build/pull/42))
+- **rsbuild-plugin-angular:** improve build speed using parallel compilation ([becac29](https://github.com/Coly010/ng-rspack-build/commit/becac29))
+- **rsbuild-plugin-angular:** handle ssr dev server changes ([7475540](https://github.com/Coly010/ng-rspack-build/commit/7475540))
+- **rsbuild-plugin-angular:** setup rsbuild environments correctly ([8265f96](https://github.com/Coly010/ng-rspack-build/commit/8265f96))
+- **rsbuild-plugin-angular:** update dev server config ([efde6fc](https://github.com/Coly010/ng-rspack-build/commit/efde6fc))
+- **rsbuild-plugin-angular:** expose express app from createServer for modification ([002d602](https://github.com/Coly010/ng-rspack-build/commit/002d602))
+- **rsbuild-plugin-angular:** correctly set ng env vars ([b81c6d8](https://github.com/Coly010/ng-rspack-build/commit/b81c6d8))
+- **rsbuild-plugin-angular:** add createServer util ([39840fd](https://github.com/Coly010/ng-rspack-build/commit/39840fd))
+- **rsbuild-plugin-angular:** add initial package ([#30](https://github.com/Coly010/ng-rspack-build/pull/30))
+
+### 🩹 Fixes
+
+- **rsbuild-plugin-angular:** ensure ssr server serves assets correctly ([0a69b55](https://github.com/Coly010/ng-rspack-build/commit/0a69b55))
+- **rsbuild-plugin-angular:** force rsbuild to use node.js module resolution ([34b6b80](https://github.com/Coly010/ng-rspack-build/commit/34b6b80))
+- **rsbuild-plugin-angular:** process all npm packages in build ([0361bc7](https://github.com/Coly010/ng-rspack-build/commit/0361bc7))
+- **rsbuild-plugin-angular:** ensure package patch is resolved correctly ([b11ae63](https://github.com/Coly010/ng-rspack-build/commit/b11ae63))
+- **rsbuild-plugin-angular:** handle ssr dev server live reloads ([481dfcb](https://github.com/Coly010/ng-rspack-build/commit/481dfcb))
+- **rsbuild-plugin-angular:** only set isServer when running in server env ([48332d1](https://github.com/Coly010/ng-rspack-build/commit/48332d1))
+
+### ❤️ Thank You
+
+- Colum Ferry @Coly010
+
 ## 19.0.0-alpha.18 (2025-01-15)
 
 ### 🩹 Fixes